isolate similarities storing in its own module
