+++ propius/__init__.py
@@ -2,14 +2,17 @@ __version__ = "0.1"
 
 from propius import utils
 from propius import model
+from propius import persistance
 
 from .utils import stream_csv
 from .model import SimilarityModel
-from .model import ModelStorer
+from .persistance import ModelStorer
 
 __all__ = [
     "utils",
     "model",
+    "persistance",
     "stream_csv",
     "SimilarityModel",
+    "ModelStorer"
 ]
